Extract parameter Theta into spiral ctor
