+++ cs/TagsCloudVisualization/Tests/ArchimedeanSprial_Should.cs
@@ -10,21 +10,23 @@ namespace TagsCloudVisualization.Tests
     [TestFixture]
     public class ArchimedeanSprial_Should
     {
+        private const double theta = 0.5;
+        
         [TestCase(0, 0, TestName = "EmptyPoint")]
         [TestCase(1, 2, TestName = "PointWithPositiveCoords")]
         [TestCase(-1, -2, TestName = "PointWithNegativeCoords")]
         [TestCase(1, -2, TestName = "PointWithMixedCoords")]
         public void ArchimedeanSpiralCtor_OnAllPoints_DoesNotThrow(int x, int y)
         {
-            Action callCtor = () => _ = new ArchimedeanSpiral(new Point(x, y));
+            Action callCtor = () => _ = new ArchimedeanSpiral(new Point(x, y), theta);
             callCtor.Should().NotThrow();
         }
 
         [Test]
         public void GetNextPoint_OnEqualSpiralInstance_ShouldBeEqual()
         {
-            var firstSpiral = new ArchimedeanSpiral(Point.Empty);
-            var secondSpiral = new ArchimedeanSpiral(Point.Empty);
+            var firstSpiral = new ArchimedeanSpiral(Point.Empty, theta);
+            var secondSpiral = new ArchimedeanSpiral(Point.Empty, theta);
 
             var firstPoints = Enumerable.Range(0, 50).Select(i => firstSpiral.GetNextPoint());
             var secondPoints = Enumerable.Range(0, 50).Select(i => secondSpiral.GetNextPoint());
@@ -34,8 +36,8 @@ namespace TagsCloudVisualization.Tests
         [Test]
         public void GetNextPoint_OnDifferentSpiralInstance_ShouldBeDifferent()
         {
-            var firstSpiral = new ArchimedeanSpiral(Point.Empty);
-            var secondSpiral = new ArchimedeanSpiral(new Point(1000, 1000));
+            var firstSpiral = new ArchimedeanSpiral(Point.Empty, theta);
+            var secondSpiral = new ArchimedeanSpiral(new Point(1000, 1000), theta);
 
             var firstPoints = Enumerable.Range(0, 50).Select(i => firstSpiral.GetNextPoint());
             var secondPoints = Enumerable.Range(0, 50).Select(i => secondSpiral.GetNextPoint());
