+++ cs/TagsCloudVisualization/Visualization/CircularCloudLayouterVisualizer.cs
@@ -4,7 +4,6 @@ using System.Drawing.Imaging;
 using System.IO;
 using System.Linq;
 using TagsCloudVisualization.Core;
-using TagsCloudVisualization.VisualizationSettings;
 
 namespace TagsCloudVisualization.Visualization
 {
